requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8
unittest-data-provider
numpy
django-bootstrap-v5
djangorestframework